Create PT1.ASS7***.11.12.23.17:06
he only change I made was to add the using namespace std; statement at the beginning of the program. This allows me to use the names of standard library objects without having to prefix them with std::.

For example, in the original program, I would have to write std::cout to print a message to the console. In the rewritten program, I can simply write cout.

Using the std namespace is generally considered to be good practice, as it makes your code more readable and concise.